backend/requirements.txt
# ==============================================================================
# Backend Dependencies - Personal Federated Learning Knowledge Base Agent
# ==============================================================================

# Flask & Core
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-CORS>=4.0.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0

# Authentication
PyJWT>=2.8.0

# Security
werkzeug>=3.0.0
cryptography>=41.0.0

# Database
SQLAlchemy>=2.0.0
chromadb>=0.4.22

# Document Parsing (Docling)
docling>=2.0.0
docling-core>=2.0.0

# Embeddings & Reranking
sentence-transformers>=2.2.0

# LLM Orchestration
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# API Clients
openai>=1.0.0
httpx>=0.27.0
dashscope>=1.14.0

# RAG Evaluation
ragas>=0.1.0
datasets>=2.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Testing
pytest>=8.0.0
pytest-flask>=1.3.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0

# Type hints
typing-extensions>=4.9.0
